requirements.txt
python-dateutil
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.0
django-filter==2.2.0
djangorestframework==3.10.3
gunicorn==19.9.0
oauth==1.0.1
oauthlib==3.1.0
Pillow==5.1.0
PyJWT==1.7.1
social-auth-app-django==3.1.0
social-auth-core==3.2.0
urllib3==1.25.7
whitenoise==4.1.4
sendgrid==6.1.1
mysqlclient==1.4.6
validate-email==1.3
certifi==2018.1.18
chardet==3.0.4
idna==2.6
pycryptodome==3.6
requests==2.18.4
integration-portal==0.1